Refactor Route53 provider to use @aws-sdk/client-route-53 and add cache utility functions
